requirements.txt
Flask==3.1.2
matplotlib==3.10.6
wordcloud==1.9.4
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.0.3
plotly==5.18.0
requests==2.31.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-dotenv==1.0.0
aiohttp==3.9.1
schedule==1.2.0